pyproject.toml
[tool.poetry]
name = "langgraphpython"
version = "0.0.1"
description = "LangGraph Python learning project"
authors = ["masashimorita"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.1"
black = "^25.1.0"
isort = "^6.0.1"
langchain = "^0.3.24"
langchain-openai = "^0.3.14"
langgraph = "^0.3.34"
grandalf = "^0.8"
langchain-tavily = "^0.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
